move dm_owner, and shut_up cmd to util
